agents/news-agent/pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "agentic-layer-sdk-adk>=0.3.2",
]
